refactoring to use a similar class
